requirements.txt
flit
numpy
openai
PyYAML
sentence-transformers
nbsphinx
